Add SWC plugin imports after "use client" directive (#9445)
When adding a "use client" directive to a module, it must come before
any other expressions. Instead of always inserting import declarations
as the first item, the SWC plugin now checks if the first expression in
a module's body is a "use client" directive and inserts its imports
after it, if that's the case.

Closes: dotansimha/graphql-code-generator#9445